+++ test/gosh_test.js
@@ -6,8 +6,7 @@ const { DocumentStore } = require('../lib/gosh')
 describe('Gosh', () => {
   it('lets you build a store and get things by a unique index', () => {
     const dave = { age: 22, uid: 'abcdef123' }
-    const makeId = document => document.uid
-    const store = new DocumentStore({ makeId })
+    const store = new DocumentStore(document => document.uid)
       .withUniqueIndex(document => document.age.toString())
       .put(dave)
     const actual = store.get({ age: 22 })
@@ -18,8 +17,7 @@ describe('Gosh', () => {
     const dave = { name: 'Dave', hair: 'red' }
     const dan = { name: 'Dan', hair: 'red' }
     const susan = { name: 'Susan', hair: 'grey' }
-    const makeId = document => document.name
-    const store = new DocumentStore({ makeId })
+    const store = new DocumentStore(document => document.name)
       .withOneToManyIndex(document => document.hair)
       .put(dave)
       .put(dan)
@@ -30,8 +28,7 @@ describe('Gosh', () => {
 
   it('implicitly adds a unique index on the Id', () => {
     const dave = { uid: 'abcdef123' }
-    const makeId = document => document.uid
-    const store = new DocumentStore({ makeId })
+    const store = new DocumentStore(document => document.uid)
       .put(dave)
     const actual = store.get({ uid: 'abcdef123' })
     assert.equal(actual, dave)