+++ CHANGELOG.md
@@ -7,12 +7,11 @@ This project adheres to [Semantic Versioning](http://semver.org).
 This document is formatted according to the principles of [Keep A CHANGELOG](http://keepachangelog.com).
 
 ----
-
 ## [In Git - targetting v2.1.0](https://github.com/mattwynne/gosh/compare/v2.0.0...master) (Not released)
 
 ### Added
 
-* Implicitly add a unique index on the ID field
+* N/A
 
 ### Changed
 
@@ -24,7 +23,29 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 
 ### Removed
 
-* `DocumentStore#find` replaced with `DocumentStore#get`
+* N/A
+
+### Fixed
+
+* N/A
+
+## [v3.0.0](https://github.com/mattwynne/gosh/compare/v2.0.0...v3.0.0)
+
+### Added
+
+* Implicitly add a unique index on the ID field
+
+### Changed
+
+* Changed constructor on DocumentStore to take `makeId` function as a regular argument
+
+### Deprecated
+
+* N/A
+
+### Removed
+
+* N/A
 
 ### Fixed